requirements.txt
azure-devops
langchain
GitPython
msrestazure